Use parse_program instead of vernacular_compile
vernacular_compile seems to produce wrong results when used with example programs in the tutorial lesson "Programs for Arithmetic". For example, try running the successor function. Evaluate and eval_slow seem to give different results on  http://rrose1.github.io/jsonesharp/. (The former, but not the latter, uses vernacular_compile.)